feat(hir,mir,runtime): align on_crash crash_code ABI to i64 and load std/failure.hew TypeDecls
Three coordinated changes unblock info.code field access in #[on(crash)]
handlers:

Runtime ABI (hew-runtime):
- HewOnCrashFn crash_code parameter c_int → i64.
- supervisor.rs call site widens with crash_code as i64.
- on_crash_invocation test helper updated to i64 / AtomicI64.

HIR TypeDecl loading (hew-hir):
- Non-root module graph walk (pre-pass + fourth pass) registers and
  emits pub TypeDecl items. PanicInfo field layout from std/failure.hew
  now reaches MIR record_field_orders when imported.

MIR prologue injection (hew-mir):
- lower_lifecycle_hooks on_crash arm replaces PanicInfo parameter with
  i64 crash_code and prepends a synthetic let info: PanicInfo =
  PanicInfo { code: __crash_code } so FieldAccess lowering finds
  info.code via record_field_orders.
- Unconditional PanicInfo registration in record_field_orders +
  is_known_actor_runtime_ty (annotated SHIM with WHEN OBSOLETE
  markers — proper fix is loading register_builtin_failure_surface
  via search paths).

Acceptance fixture: on_crash_info_code.hew compiles and exits 42
through the full HIR→MIR→codegen pipeline.

+++ hew-runtime/src/internal/types.rs
@@ -1,7 +1,5 @@
 //! Internal type definitions shared across runtime modules.
 
-use std::ffi::c_int;
-
 use crate::execution_context::HewExecutionContext;
 
 /// Actor dispatch function signature (context-leading canonical).
@@ -21,10 +19,16 @@ pub type HewDispatchFn = unsafe extern "C-unwind" fn(
 /// policy is applied. Receives the execution context, the crash code
 /// (trap kind integer), and the actor's current state pointer.
 ///
-/// `void (*on_crash)(HewExecutionContext *ctx, int crash_code, void *actor_state_ptr)`
+/// `crash_code` is i64 to match the `code: i64` field of `PanicInfo` in
+/// `std/failure.hew`. The supervisor's internal plumbing tracks the code as
+/// `c_int` and widens it to `i64` at the call site so the internal event
+/// struct and the public C ABI (`hew_supervisor_notify_child_event`) stay
+/// unchanged.
+///
+/// `void (*on_crash)(HewExecutionContext *ctx, int64_t crash_code, void *actor_state_ptr)`
 pub type HewOnCrashFn = unsafe extern "C" fn(
     ctx: *mut HewExecutionContext,
-    crash_code: c_int,
+    crash_code: i64,
     actor_state_ptr: *mut std::ffi::c_void,
 );
